# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "User login",
        "description": "Authenticates a user and returns a JWT token along with user information.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials or account deactivated"
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "User registration",
        "description": "Creates a new user account and returns a JWT token along with user information.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registration successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error - invalid input data"
          },
          "409": {
            "description": "User with email or username already exists"
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/profile": {
      "get": {
        "operationId": "AuthController_getProfile",
        "summary": "Get current user profile",
        "description": "Returns the profile information of the currently authenticated user.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Profile retrieved successfully"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          }
        },
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refreshToken",
        "summary": "Refresh access token",
        "description": "Validates the refresh token and returns a new access token. Implements sliding expiration: each refresh extends the refresh token by 60 days. After 60 days of complete inactivity, the refresh token expires and user must login again.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Refresh token received from login/register",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed successfully. Returns new access token and same refresh token (with extended expiration).",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                    "refreshToken": "a1b2c3d4e5f6g7h8i9j0..."
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired refresh token"
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/status": {
      "get": {
        "operationId": "AuthController_checkStatus",
        "summary": "Check authentication status",
        "description": "Verifies if the current token is valid and returns authentication status.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Authentication status verified"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          }
        },
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/auth/google/callback": {
      "post": {
        "operationId": "AuthController_googleAuthCallback",
        "summary": "Google OAuth2 - Validate ID Token",
        "description": "Validates a Google ID Token (credential) received from Google Identity Services SDK. Frontend should use Google's JavaScript library to get the id_token and send it here.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Google ID Token (credential) from Google Identity Services",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleAuthCodeDto"
              },
              "examples": {
                "example1": {
                  "summary": "ID Token from Google Identity Services",
                  "value": {
                    "id_token": "eyJhbGciOiJSUzI1NiIsImtpZCI6IjI3..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authorization code exchanged successfully",
            "schema": {
              "example": {
                "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                "provider": "google",
                "isNewUser": false,
                "user": {
                  "id": "123e4567-e89b-12d3-a456-426614174000",
                  "email": "user@gmail.com",
                  "username": "google_1234567890",
                  "firstName": "John",
                  "lastName": "Doe",
                  "avatar": "https://lh3.googleusercontent.com/a/ACg8ocJ...",
                  "role": "USER",
                  "isActive": true
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoogleAuthResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid authorization code or Google authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Invalid authorization code",
                    "error": "INVALID_AUTHORIZATION_CODE",
                    "details": "The authorization code is invalid, expired, or has already been used. Please try signing in again."
                  }
                }
              }
            }
          },
          "401": {
            "description": "Google email not verified or account deactivated",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Google email is not verified",
                    "error": "EMAIL_NOT_VERIFIED",
                    "details": "Please verify your email address with Google before signing in."
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/api/v1/articles": {
      "post": {
        "operationId": "ArticlesController_create",
        "summary": "Create a new article",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateArticleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Articles"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "ArticlesController_findAll",
        "summary": "Get all articles",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Articles"
        ]
      }
    },
    "/api/v1/articles/{id}": {
      "get": {
        "operationId": "ArticlesController_findOne",
        "summary": "Get an article by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Articles"
        ]
      },
      "patch": {
        "operationId": "ArticlesController_update",
        "summary": "Update an article",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateArticleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Articles"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ArticlesController_remove",
        "summary": "Delete an article",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Articles"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/articles/{id}/propose-changes": {
      "post": {
        "operationId": "ArticlesController_proposeChanges",
        "summary": "Propose changes to an article",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateArticleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Articles"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "♻️ Recycling Centers Management API",
    "description": "Enterprise-grade API for managing recycling centers, materials, and user operations",
    "version": "1.0",
    "contact": {
      "name": "The Wealthiest Programmer",
      "url": "https://github.com/wealthiest-programmer",
      "email": "contact@wealthiest-programmer.dev"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "john.doe@example.com"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "securePassword123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT access token (short-lived, 15min)"
          },
          "refreshToken": {
            "type": "string",
            "description": "Refresh token (long-lived, 60 days with sliding expiration)"
          },
          "user": {
            "type": "object",
            "description": "User information"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "user"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "john.doe@example.com"
          },
          "username": {
            "type": "string",
            "description": "Unique username",
            "example": "johndoe123"
          },
          "password": {
            "type": "string",
            "description": "User password (minimum 6 characters)",
            "example": "securePassword123",
            "minLength": 6
          },
          "firstName": {
            "type": "string",
            "description": "User first name",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "description": "User last name",
            "example": "Doe"
          }
        },
        "required": [
          "email",
          "username",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Refresh token to exchange for new access token",
            "example": "a1b2c3d4e5f6g7h8i9j0..."
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "GoogleAuthCodeDto": {
        "type": "object",
        "properties": {
          "id_token": {
            "type": "string",
            "description": "Google ID Token (JWT) received from Google Identity Services",
            "example": "eyJhbGciOiJSUzI1NiIsImtpZCI6IjI3..."
          }
        },
        "required": [
          "id_token"
        ]
      },
      "GoogleAuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT access token (short-lived, 15min)"
          },
          "refreshToken": {
            "type": "string",
            "description": "Refresh token (long-lived, 60 days with sliding expiration)"
          },
          "user": {
            "type": "object",
            "description": "User information"
          },
          "provider": {
            "type": "string",
            "description": "Google authentication provider",
            "example": "google"
          },
          "isNewUser": {
            "type": "boolean",
            "description": "Whether this is a new user registration",
            "example": false
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "user",
          "provider",
          "isNewUser"
        ]
      },
...
      "CreateArticleDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the article"
          },
          "slug": {
            "type": "string",
            "description": "Slug of the article"
          },
          "content": {
            "type": "string",
            "description": "Content of the article"
          },
          "summary": {
            "type": "string",
            "description": "Summary"
          },
          "tags": {
            "description": "Tags",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title",
          "slug",
          "content"
        ]
      },
      "UpdateArticleDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the article"
          },
          "slug": {
            "type": "string",
            "description": "Slug of the article"
          },
          "content": {
            "type": "string",
            "description": "Content of the article"
          },
          "summary": {
            "type": "string",
            "description": "Summary"
          },
          "tags": {
            "description": "Tags",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title",
          "slug",
          "content"
        ]
      },
...
    }
  }
}